Add trait bound

+++ src/addr_modes/indirect_x.rs
@@ -8,8 +8,8 @@ impl AddressingMode for IndirectX {
         let zero_page_offset = ctx.read(ctx.pc) + ctx.x;
         ctx.pc += 1;
 
-        let lo = ctx.read(zero_page_offset as u16);
-        let hi = ctx.read((zero_page_offset + 1) as u16);
+        let lo = ctx.read(zero_page_offset);
+        let hi = ctx.read(zero_page_offset + 1);
 
         u16::from_le_bytes([lo, hi])
     }
